불필요한 import와 변수들 제거
