+++ src/components/VoteCards/RookieVoteCard/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState, useEffect } from 'react';
 import { Typography, Card, CardContent, Divider, CircularProgress } from '@material-ui/core';
 import styles from './styles';
-import { EpisodeVotePolls, SelectVotePolls } from '../..';
+import { SelectVotePolls } from '../..';
 import axios from 'axios';
 
 interface RookieVoteCardProps {
