+++ src/polyfills/assign.ts
@@ -3,7 +3,6 @@ export const polyfill = () => {
         // Must be writable: true, enumerable: false, configurable: true
         Object.defineProperty(Object, "assign", {
             value: function assign(target: any, varArgs: any) { // .length of function is 2
-                'use strict';
                 if (target == null) { // TypeError if undefined or null
                     throw new TypeError('Cannot convert undefined or null to object');
                 }
